Superseeds #8898 

Changes are:
* cherry-picked Seangchan's change from 8898, replacing _rev by _revision
* don't upload all docs in a single HTTP request, bu create small chunks of 100 docs per request (performance varies A LOT, from a few seconds to a few minutes)
* no need for the try/except block, all docs should be well formatted in a new agent
* creates a random revision id for the first time a document is created (to avoid issues if a doc gets deleted in central wmstats)

It appears to have a bad performance, we should keep an eye on this mode of upload and make further changes if needed.