@ticoann 2018-11-07
This doesn't seem to work.  It seems for the initial doc, _rev has to be set