@ticoann 2018-11-08
Alan, how was the performance test went. (You said initial call take long time 18 sec for 32 requests)