@ticoann 2018-11-08
This is for only new inserts, right? After the first insert is it faster (for updating records)?