@amaltaro 2018-11-08
Yes, new inserts take quite a while. Only document updates are much faster.