@amaltaro 2018-11-09
Now that we have wiped out the wmstats database, AnalyticsDataCollector became 10x faster. 
Checking a CERN agent (0250), time to upload agent_request docs went down from ~250secs to 16secs (310 docs).
While at FNAL it went down from 250secs to 50secs (~430 docs).