@amaltaro 2018-11-09
That was the execution time of the component cycle in general, so including everything. Still, I made an average (by eye) of the last 50 cycles or so.
Yes, DB is much healthier since we applied this patch yesterday, it grew only 13GB over the night.