@ticoann 2018-11-09
13G? Before the compaction? Right now it seems it is 2.7G. I expect it keeps growing until we hit the revision limit. current revision is around ~30 revision.