@amaltaro 2018-11-09
No no, it went from 332GB (yesterday afternoon) to 345GB (this morning). While the night before it grew ~150GB :)