@ticoann 2018-11-09
Thanks Alan, it looks a lot better than before. Hope it stays that way. :-)